feat: Add flag to skip unknown fields instead of error
This is mostly required for forward compatibility, when new fields
are added, any old instance will fail to deserialize a message
serialized with a newer version. This is also aligned with prost that
skips any unknown field too.

I added a flag to to the builder, so this behavior is opt-in, keeping
the old behavior (error out for unknown fields) as default.