fix: properly handle deserialization from Reader
The pattern "let s: &str = Deserialize::deserialize(deser)" used
in multiple places is not ideal, as it generates errors when the
deserializer uses an IO Reader and not a string. To fix this,
implementing a visitor is preferred, as it gives the deserializer the
choice of allocating or not a string.

To ensure the issue is fixed, the tests now deserialize both from a
string and from a reader.

Closes #55