fix: remove dependency on protoc in pbjson-types
This change moves pbjson-type's code generation to a manual executable
instead of build.rs. Now the generated files are checked into the repo,
and users do not need `protoc` to build the pbjson-types crate.

This has a slight drawback in that pbjson developers will need to invoke
the generator manually after any relevant code-gen changes. This should
generally be a worthwhile tradeoff, as several users have encountered
trouble with requiring protoc (see #62).

That drawback could potentially be averted with some CI steps to run the
generator and check-in the code automatically; this change does not
include such automation however.